[QOL-9160] merge changes from upstream and clean up
- put ckanext-report URL in dev-requirements.txt instead of having specific GitHub Actions steps
- move main plugin module and controllers to top level
- adjust plugin interface functions to support CKAN 2.10
- clean up 'six' imports and use of toolkit API
- move common controller and CLI functions into 'utils' module for compatibility with upstream
- drop obsolete Tablesorter class from reports
- add more unit tests
- use pytest to clean up assertions
