+++ .flake8
@@ -3,7 +3,6 @@
 
 exclude =
     ckan
-    scripts
     .git
 
 # Extended output format.
