--- bin/travis-flake.sh
@@ -1,3 +1,0 @@
-
-pip install -U pip wheel
-pip install -r flake-requirements.txt
