+++ ckanext/archiver/cli.py
@@ -1,10 +1,13 @@
 # encoding: utf-8
 
 import click
-from . import ArchivalCommands
+from . import utils
 
 
 # Click commands for CKAN 2.9 and above
+
+def get_commands():
+    return [archiver]
 
 
 @click.group(short_help='Download and save copies of all package resources.')
@@ -18,34 +21,45 @@ def archiver():
 def init():
     '''Creates the database table archiver needs to run.
     '''
-    ArchivalCommands().initdb()
+    utils.initdb()
+    click.secho("Archiver tables are initialized", fg="green")
 
 
 @archiver.command()
 def migrate():
     '''Updates the database schema to include new fields.
     '''
-    ArchivalCommands().migrate()
+    utils.migrate()
 
 
 @archiver.command()
 @click.option('-q', '--queue', help='Send to a particular queue')
-@click.argument(u'dataset-spec', nargs=-1)
-def update(dataset_spec, queue=None):
+@click.argument('identifiers', nargs=-1)
+def update(identifiers, queue=None):
     '''Archive all resources or just those belonging to a specific
     package or group, if specified
     '''
-    ArchivalCommands().update(dataset_spec, queue)
+    utils.update(identifiers, queue)
 
 
 @archiver.command()
 @click.option('-q', '--queue', help='Send to a particular queue')
-@click.argument(u'dataset-spec', nargs=-1)
-def update_test(dataset_spec, queue=None):
+@click.argument('identifiers', nargs=-1)
+def update_test(identifiers, queue=None):
     '''Does an archive in the current process i.e. avoiding worker queue
     so that you can test on the command-line more easily.
     '''
-    ArchivalCommands(queue).update_test(dataset_spec)
+    utils.update_test(identifiers, queue)
+    click.secho("Completed test update", fg="green")
+
+
+@archiver.command()
+@click.argument('package_ref', required=False)
+def view(package_ref):
+    '''Views archival info, in general and if you specify one, about
+    a particular dataset's resources.
+    '''
+    utils.view(package_ref)
 
 
 @archiver.command()
@@ -54,23 +68,14 @@ def clean_status():
      archived resource, whether it was successful or not, with errors.
      It does not change the cache_url etc. in the Resource
     '''
-    ArchivalCommands().clean_status()
+    utils.clean_status()
 
 
 @archiver.command()
 def clean_cached_resources():
     '''Removes all cache_urls and other references to resource files on disk.
     '''
-    ArchivalCommands().clean_cached_resources()
-
-
-@archiver.command()
-@click.argument(u'dataset_spec')
-def view(dataset_spec):
-    '''Views archival info, in general and if you specify one, about
-    a particular dataset's resources.
-    '''
-    ArchivalCommands().view(dataset_spec)
+    utils.clean_cached_resources()
 
 
 @archiver.command()
@@ -81,7 +86,7 @@ def report(outputfile):
     orphan. The outputfile parameter is the name of the CSV output
     from running the report.
     '''
-    ArchivalCommands.report(outputfile, False)
+    utils.report(outputfile, False)
 
 
 @archiver.command()
@@ -93,7 +98,7 @@ def delete_orphans(outputfile):
     Does not cleanup IUploader plugin uploaded files as they may not be
     on disk.
     '''
-    ArchivalCommands().report(outputfile, True)
+    utils.report(outputfile, True)
 
 
 @archiver.command()
@@ -105,14 +110,14 @@ def migrate_archive_dirs():
     Running this moves them to the new locations and updates the
     cache_url on each resource to reflect the new location.
     '''
-    ArchivalCommands().migrate_archive_dirs()
+    utils.migrate_archive_dirs()
 
 
 @archiver.command()
 def size_report():
     '''Reports on the sizes of files archived.
     '''
-    ArchivalCommands().size_report()
+    utils.size_report()
 
 
 @archiver.command()
@@ -121,8 +126,4 @@ def delete_files_larger_than_max():
     want to delete archived files that are now above the threshold,
     and stop referring to these files in the Archival table of the db.
     '''
-    ArchivalCommands().delete_files_larger_than_max_content_length()
-
-
-def get_commands():
-    return [archiver]
+    utils.delete_files_larger_than_max_content_length()
