+++ ckanext/archiver/reports.py
@@ -2,7 +2,6 @@
 
 import copy
 import json
-import six
 try:
     from collections import OrderedDict  # from python 2.7
 except ImportError:
@@ -91,7 +90,7 @@ def broken_links_index(include_sub_organizations=False):
     num_broken_resources = 0
     num_packages = 0
     num_resources = 0
-    for org_name, org_counts in six.iteritems(results):
+    for org_name, org_counts in results.items():
         data.append(OrderedDict((
             ('organization_title', results[org_name]['organization_title']),
             ('organization_name', org_name),
