+++ ckanext/archiver/tasks.py
@@ -8,25 +8,30 @@ import mimetypes
 import os
 import requests
 import shutil
-from six.moves.urllib import parse as urlparse
 import six
+from six import text_type as str
+from six.moves import http_client
+from six.moves.urllib.parse import urlparse, urljoin, quote, urlunparse
 import tempfile
-import time
-import urllib
+from time import sleep
 
 from requests.packages import urllib3
 
-from ckan import plugins as p
+from ckan import model, plugins as p
 from ckan.common import _
 from ckan.lib import uploader
+from ckan.lib.search.index import PackageSearchIndex
 from ckan.lib.uploader import ResourceUpload as DefaultResourceUpload
-from ckanext.archiver import interfaces as archiver_interfaces
-from ckanext.archiver import default_settings as settings
+from ckan.plugins import toolkit
+from ckan.plugins.toolkit import config
+
+from . import interfaces as archiver_interfaces, \
+    default_settings as settings
+from .model import Status, Archival
+from .requests_ssl import SSLv3Adapter
 
 import logging
 
-toolkit = p.toolkit
-config = toolkit.config
 log = logging.getLogger(__name__)
 
 ALLOWED_SCHEMES = set(('http', 'https', 'ftp'))
@@ -111,14 +116,15 @@ class CkanError(ArchiverError):
     pass
 
 
-def update_resource(ckan_ini_filepath=None, resource_id=None, queue='bulk'):
+def update_resource(resource_id=None, queue='bulk'):
     '''
     Archive a resource.
     '''
+
     log.info('Starting update_resource task: res_id=%r queue=%s', resource_id, queue)
 
     # HACK because of race condition #1481
-    time.sleep(2)
+    sleep(2)
 
     # Do all work in a sub-routine since it can then be tested without celery.
     # Also put try/except around it is easier to monitor ckan's log rather than
@@ -135,10 +141,11 @@ def update_resource(ckan_ini_filepath=None, resource_id=None, queue='bulk'):
         raise
 
 
-def update_package(ckan_ini_filepath=None, package_id=None, queue='bulk'):
+def update_package(package_id=None, queue='bulk'):
     '''
     Archive a package.
     '''
+
     log.info('Starting update_package task: package_id=%r queue=%s',
              package_id, queue)
 
@@ -158,8 +165,6 @@ def update_package(ckan_ini_filepath=None, package_id=None, queue='bulk'):
 
 
 def _update_package(package_id, queue, log):
-    from ckan import model
-
     get_action = toolkit.get_action
 
     num_archived = 0
@@ -191,8 +196,6 @@ def _update_search_index(package_id, log):
     '''
     Tells CKAN to update its search index for a given package.
     '''
-    from ckan import model
-    from ckan.lib.search.index import PackageSearchIndex
     package_index = PackageSearchIndex()
     context_ = {'model': model, 'ignore_auth': True, 'session': model.Session,
                 'use_cache': False, 'validate': False}
@@ -223,8 +226,6 @@ def _update_resource(resource_id, queue, log):
         }
     If not successful, returns None.
     """
-    from ckan import model
-    from ckanext.archiver.model import Status, Archival
 
     get_action = toolkit.get_action
 
@@ -259,7 +260,7 @@ def _update_resource(resource_id, queue, log):
         upload = uploader.get_resource_uploader(resource)
         filepath = upload.get_path(resource['id'])
 
-        hosted_externally = not url.startswith(config['ckan.site_url']) or urlparse.urlparse(filepath).scheme != ''
+        hosted_externally = not url.startswith(config['ckan.site_url']) or urlparse(filepath).scheme != ''
         # if resource.get('resource_type') == 'file.upload' and not hosted_externally:
         if not hosted_externally:
             log.info("Won't attempt to archive resource uploaded locally: %s", resource['url'])
@@ -322,26 +323,27 @@ def _update_resource(resource_id, queue, log):
         'cache_url_root': config.get('ckanext-archiver.cache_url_root'),
         'previous': Archival.get_for_resource(resource_id)
     }
-    error = {'args': ''}
+
+    err = None
     try:
         download_result = download(context, resource)
     except NotChanged as e:
-        error = e
         download_status_id = Status.by_text('Content has not changed')
         try_as_api = False
         requires_archive = False
+        err = e
     except LinkInvalidError as e:
-        error = e
         download_status_id = Status.by_text('URL invalid')
         try_as_api = False
+        err = e
     except (DownloadException, DownloadError) as e:
-        error = e
         download_status_id = Status.by_text('Download error')
         try_as_api = True
+        err = e
     except ChooseNotToDownload as e:
-        error = e
         download_status_id = Status.by_text('Chose not to download')
         try_as_api = False
+        err = e
     except Exception as e:
         if os.environ.get('DEBUG'):
             raise
@@ -349,9 +351,9 @@ def _update_resource(resource_id, queue, log):
         _save(Status.by_text('Download failure'), e, resource)
         return
 
-    if not Status.is_ok(download_status_id):
+    if not Status.is_ok(download_status_id) and err:
         log.info('GET error: %s - %r, %r "%s"',
-                 Status.by_id(download_status_id), error, error.args,
+                 Status.by_id(download_status_id), err, err.args,
                  resource.get('url'))
 
         if try_as_api:
@@ -362,8 +364,8 @@ def _update_resource(resource_id, queue, log):
             # from the previous download (i.e. not when we tried it as an API)
 
         if not try_as_api or not Status.is_ok(download_status_id):
-            extra_args = [error.url_redirected_to] if hasattr(error, 'url_redirected_to') else []
-            _save(download_status_id, error, resource, *extra_args)
+            extra_args = [err.args.url_redirected_to] if 'url_redirected_to' in err.args else []
+            _save(download_status_id, err, resource, *extra_args)
             return
 
     if not requires_archive:
@@ -411,7 +413,6 @@ def download(context, resource, url_timeout=30,
     Returns a dict of results of a successful download:
       mimetype, size, hash, headers, saved_file, url_redirected_to
     '''
-    from ckanext.archiver import default_settings as settings
 
     if max_content_length == 'default':
         max_content_length = settings.MAX_CONTENT_LENGTH
@@ -511,7 +512,7 @@ def download(context, resource, url_timeout=30,
     try:
         length, hash, saved_file_path = _save_resource(resource, res, max_content_length)
     except ChooseNotToDownload as e:
-        raise ChooseNotToDownload(six.text_type(e), url_redirected_to)
+        raise ChooseNotToDownload(str(e), url_redirected_to)
     log.info('Resource saved. Length: %s File: %s', length, saved_file_path)
 
     # zero length (or just one byte) indicates a problem
@@ -547,7 +548,7 @@ def _file_hashnlength(local_path):
 
             buf = afile.read(BLOCKSIZE)
 
-    return (six.text_type(hasher.hexdigest()), length)
+    return (str(hasher.hexdigest()), length)
 
 
 def archive_resource(context, resource, log, result=None, url_timeout=30):
@@ -572,17 +573,16 @@ def archive_resource(context, resource, log, result=None, url_timeout=30):
     # my_storage_path/archive/16/165900ba-3c60-43c5-9e9c-9f8acd0aa93f/data.csv
     relative_archive_path = os.path.join(resource['id'][:2], resource['id'])
     if not uploaderHasDownloadEnabled:
-        from ckanext.archiver import default_settings as settings
         archive_dir = os.path.join(settings.ARCHIVE_DIR, relative_archive_path)
         if not os.path.exists(archive_dir):
             os.makedirs(archive_dir)
     # try to get a file name from the url
-    parsed_url = urlparse.urlparse(resource.get('url'))
+    parsed_url = urlparse(resource.get('url'))
     try:
         file_name = parsed_url.path.split('/')[-1] or 'resource'
         file_name = file_name.strip()  # trailing spaces cause problems
         file_name = file_name.encode('ascii', 'ignore')  # e.g. u'\xa3' signs
-        file_name = six.text_type(file_name)
+        file_name = str(file_name)
     except Exception:
         file_name = "resource"
 
@@ -632,8 +632,8 @@ def archive_resource(context, resource, log, result=None, url_timeout=30):
             log.warning('Not saved cache_url because no value for '
                         'ckanext-archiver.cache_url_root in config')
             raise ArchiveError(_('No value for ckanext-archiver.cache_url_root in config'))
-        cache_url = urlparse.urljoin(context['cache_url_root'],
-                                     '%s/%s' % (relative_archive_path, file_name))
+        cache_url = urljoin(str(context['cache_url_root']),
+                            '%s/%s' % (str(relative_archive_path), str(file_name)))
         return {'cache_filepath': saved_file,
                 'cache_url': cache_url}
 
@@ -682,7 +682,6 @@ def _set_user_agent_string(headers):
     Update the passed headers object with a `User-Agent` key, if there is a
     USER_AGENT_STRING option in settings.
     '''
-
     ua_str = settings.USER_AGENT_STRING
     if ua_str is not None:
         headers['User-Agent'] = ua_str
@@ -703,10 +702,10 @@ def tidy_url(url):
         try:
             url = url.decode('ascii')
         except Exception:
-            parts = list(urlparse.urlparse(url))
-            parts[2] = urllib.quote(parts[2].encode('utf-8'))
-            parts[1] = urllib.quote(parts[1].encode('utf-8'))
-            url = urlparse.urlunparse(parts)
+            parts = list(urlparse(url))
+            parts[2] = quote(parts[2].encode('utf-8'))
+            parts[1] = quote(parts[1].encode('utf-8'))
+            url = urlunparse(parts)
     url = six.text_type(url)
 
     # strip whitespace from url
@@ -772,9 +771,6 @@ def save_archival(resource, status_id, reason, url_redirected_to,
     May propagate a CkanError.
     '''
     now = datetime.datetime.utcnow()
-
-    from ckanext.archiver.model import Archival, Status
-    from ckan import model
 
     archival = Archival.get_for_resource(resource['id'])
     first_archival = not archival
@@ -842,7 +838,18 @@ def requests_wrapper(log, func, *args, **kwargs):
         res = requests.get(url, timeout=url_timeout)
     '''
     try:
-        response = func(*args, **kwargs)
+        try:
+            response = func(*args, **kwargs)
+        except requests.exceptions.ConnectionError as e:
+            if 'SSL23_GET_SERVER_HELLO' not in str(e):
+                raise
+            log.info('SSLv23 failed so trying again using SSLv3: %r', args)
+            requests_session = requests.Session()
+            requests_session.mount('https://', SSLv3Adapter())
+            func = {requests.get: requests_session.get,
+                    requests.post: requests_session.post}[func]
+            response = func(*args, **kwargs)
+
     except requests.exceptions.ConnectionError as e:
         raise DownloadException(_('Connection error: %s') % e)
     except requests.exceptions.HTTPError as e:
@@ -978,7 +985,7 @@ def link_checker(context, data):
     try:
         res = requests.head(url, timeout=url_timeout)
         headers = res.headers
-    except requests.exceptions.InvalidURL as ve:
+    except http_client.InvalidURL as ve:
         log.error("Could not make a head request to %r, error is: %s."
                   " Package is: %r. This sometimes happens when using an old version of requests on a URL"
                   " which issues a 301 redirect. Version=%s", url, ve, data.get('package'), requests.__version__)
