+++ ckanext/archiver/tests/data/wms_getcap_1.3.xml
