+++ conftest.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+pytest_plugins = [
+    u'ckanext.archiver.tests.fixtures'
+]
