+++ dev-requirements.txt
@@ -1,6 +1,3 @@
-ckanapi==4.3
-factory-boy==2.1.1
-flask==1.0.2
 mock
 pytest-ckan
 pytest-cov
