--- pip-requirements.txt
@@ -1,8 +1,0 @@
-ckantoolkit==0.0.4
-pika>=1.1.0,<1.3.0
-pyOpenSSL==18.0.0
-redis
-requests>=2.11.1
-six>=1.12.0
-messytables==0.15.2
-Unidecode==1.0.22
