+++ requirements.txt
@@ -4,6 +4,6 @@ progressbar2==3.55.0
 #SQLAlchemy>=0.6.6
 #requests>=1.1.0
 ckantoolkit>=0.0.4
-six>=1.0.0
+six>=1.12.0
 
-ckanext-report
+-e git+https://github.com/qld-gov-au/ckanext-report.git@develop#egg=ckanext-report
