--- test-core.ini
@@ -1,57 +1,0 @@
-#
-# ckan - Pylons testing environment configuration
-#
-# The %(here)s variable will be replaced with the parent directory of this file
-#
-[DEFAULT]
-debug = true
-# Uncomment and replace with the address which should receive any error reports
-#email_to = you@yourdomain.com
-smtp_server = localhost
-error_email_from = paste@localhost
-
-ckan.storage_path = /tmp/storage
-
-[server:main]
-use = egg:Paste#http
-host = 0.0.0.0
-port = 5000
-
-[app:celery]
-CELERY_ALWAYS_EAGER = True
-
-[app:main]
-use = config:../ckan/test-core.ini
-
-# Logging configuration
-[loggers]
-keys = root, ckan, sqlalchemy
-
-[handlers]
-keys = console
-
-[formatters]
-keys = generic
-
-[logger_root]
-level = WARN
-handlers = console
-
-[logger_ckan]
-qualname = ckan
-handlers =
-level = INFO
-
-[logger_sqlalchemy]
-handlers =
-qualname = sqlalchemy.engine
-level = WARN
-
-[handler_console]
-class = StreamHandler
-args = (sys.stdout,)
-level = NOTSET
-formatter = generic
-
-[formatter_generic]
-format = %(asctime)s %(levelname)-5.5s [%(name)s] %(message)s
