+++ test.ini
@@ -3,9 +3,6 @@ debug = false
 smtp_server = localhost
 error_email_from = paste@localhost
 
-ckan.storage_path = /tmp/storage
-
-
 [server:main]
 use = egg:Paste#http
 host = 0.0.0.0
@@ -15,16 +12,12 @@ port = 5000
 CELERY_ALWAYS_EAGER = True
 
 [app:main]
-use = config:test-core.ini
-# Here we hard-code the database and a flag to make default tests
-# run fast.
-sqlalchemy.url = sqlite:///
+use = config:../ckan/test-core.ini
+solr_url = http://127.0.0.1:8983/solr
 
-ckan.plugins = archiver report testipipe
+ckan.plugins = archiver testipipe report
 ckanext-archiver.cache_url_root = http://localhost:50001/resources/
 ckanext-archiver.max_content_length = 1000000
-# NB: other test configuration should go in test-core.ini, which is
-#     what the postgres tests use.
 
 # Logging configuration
 [loggers]
@@ -43,7 +36,7 @@ handlers = console
 [logger_ckan]
 qualname = ckan
 handlers = console
-level = DEBUG
+level = INFO
 propagate = 0
 
 [logger_ckanext]
@@ -55,7 +48,7 @@ propagate = 0
 [logger_sqlalchemy]
 handlers = console
 qualname = sqlalchemy.engine
-level = WARN
+level = WARNING
 
 [logger_activity]
 level = WARNING
